Ajout de Try/Catch dans les transactions + Ajout de la procedure pour les upgrade du plugin.

+++ edit_author.php
@@ -17,9 +17,9 @@ $PAGE->set_url($CFG->wwwroot.'/mod/randomstrayquotes/edit_author.php');
 //$course_id = required_param('courseid', PARAM_INT);
 
 if (isset($_GET['authorid'])){
-   $category_id = required_param('authorid', PARAM_INT);
+   $authorid = required_param('authorid', PARAM_INT);
  }else{
-   $category_id = $_POST['authorid'];
+   $authorid = $_POST['authorid'];
 }
 
 if (isset($_GET['courseid'])){
@@ -55,38 +55,62 @@ if ($form->is_cancelled()) {
     redirect(new moodle_url('/mod/randomstrayquotes/list_authors.php', ['courseid' => $course_id,  'userid' => $USER->id ]));
 }
 
+if (isset($_REQUEST['backtolist'])) {
+    redirect(new moodle_url('/mod/randomstrayquotes/list_authors.php', ['courseid' => $courseid,  'userid' => $USER->id ]));
+}
+
 if (isset($_REQUEST['delete'])) {
   try{
       $transaction = $DB->start_delegated_transaction();
+      // We delete the author
       $table = 'randomstrayquotes_authors';
       $DB->delete_records($table, array('id' => $_POST['authorid']));
-      $url= new moodle_url('/mod/randomstrayquotes/list_authors.php', array('id' => $courseid));
+      // We delete the quotes associated with this author
+      $table2 = 'randomstrayquotes_quotes';
+      $DB->delete_records($table2, array('author_id' => $_POST['authorid']));
+
+      $url= new moodle_url('/mod/randomstrayquotes/list_authors.php', array('courseid' => $courseid));
       $transaction->allow_commit();
       }
        catch (\Exception $e) {
              $transaction->rollback($e);
-             $url= new moodle_url('/mod/randomstrayquotes/edit_author.php', array('id' => $courseid, 'status' =>'error'));
+             $url= new moodle_url('/mod/randomstrayquotes/edit_author.php', array('courseid' => $courseid, 'status' =>'error'));
              redirect($url, 'Some error have occured', 3);
       }
       redirect($url, 'Transaction successful', 3);
 }
 
+// If the is data in the submitted form we update the record
+if ($data = $form->get_data()) {
+
+  try{
+      $transaction = $DB->start_delegated_transaction();
+      // ... store or update $entry
+      file_save_draft_area_files($data->userfile, $ctx->id, 'mod_randomstrayquotes', 'content',
+                     $data->userfile, array('subdirs' => 0, 'maxbytes' => $maxbytes, 'maxfiles' => 50));
+      // We obtain de data about the author
+      $author = $DB->get_record('randomstrayquotes_authors', ['id' => $data->authorid]);
+      // We change the author's associated picture with the one we just upoalded in the file manager
+      $author->author_picture = $data->userfile;
+      $author->author_name = $data->author_name;
+
+      // We update the author's informations with the new picture
+      $DB->update_record('randomstrayquotes_authors', $author);
+
+      $url= new moodle_url('/mod/randomstrayquotes/edit_author.php', array('courseid' => $courseid, 'authorid'=>$data->authorid));
+      $transaction->allow_commit();
+      }
+       catch (\Exception $e) {
+             $transaction->rollback($e);
+             $url= new moodle_url('/mod/randomstrayquotes/list_author.php', array('courseid' => $courseid, 'status' =>'error'));
+             redirect($url, 'Some error have occured', 3);
+      }
+      redirect($url, 'Transaction successful', 3);
+
+}
 // Creation of the page
 echo $OUTPUT->header();
 echo $OUTPUT->heading('Edit an author\'s informations');
-// If the is data in the submitted form we update the record
-if ($data = $form->get_data()) {
-    // ... store or update $entry
-    file_save_draft_area_files($data->userfile, $ctx->id, 'mod_randomstrayquotes', 'content',
-                   $data->userfile, array('subdirs' => 0, 'maxbytes' => $maxbytes, 'maxfiles' => 50));
-    // We obtain de data about the author
-    $author = $DB->get_record('randomstrayquotes_authors', ['id' => $data->authorid]);
-    // We change the author's associated picture with the one we just upoalded in the file manager
-    $author->author_picture = $data->userfile;
-
-    // We update the author's informations with the new picture
-    $DB->update_record('randomstrayquotes_authors', $author);
-}
 // Display the form
 echo $form->display();
 //Display the footer
